+++ tests/project_list.rs
@@ -45,7 +45,7 @@ project2
     let mut cmd = Command::cargo_bin("ebb")?;
     cmd.arg("project")
         .arg("list")
-        .env("EBB_CONFIG_DIR", tmp.path())
+        .env("EBB_DATA_DIR", tmp.path())
         .assert()
         .success()
         .stdout(expected_output);