+++ tests/vacation_add.rs
@@ -18,7 +18,7 @@ fn add_vacation_creates_file() -> Result<(), Box<dyn std::error::Error>> {
         .arg("add")
         .arg("2025-05-28")
         .arg("Mountain Day")
-        .env("EBB_CONFIG_DIR", tmp.path())
+        .env("EBB_DATA_DIR", tmp.path())
         .assert()
         .success();
 
@@ -53,7 +53,7 @@ fn add_vacation_fails_if_date_exists() -> Result<(), Box<dyn std::error::Error>>
         .arg("add")
         .arg("2025-05-28")
         .arg("Mountain Day")
-        .env("EBB_CONFIG_DIR", tmp.path());
+        .env("EBB_DATA_DIR", tmp.path());
 
     cmd.assert().failure().stderr(contains("already exists"));